File class can now write portions of a file.
It updates the buffers, it writes offsets currently not held in the buffer directly while update the buffer for delayed writing for offsets in memory.

+++ libnrcore/memory/StringList.cpp
@@ -54,7 +54,7 @@ namespace nrcore {
     }
 
     StringList::~StringList() {
-    	int len = list.length();
+    	int len = (int)list.length();
         for (int i=0; i<len; i++)
         	delete list[i];
     }
